+++ packages/layer-composer/src/generators/carto-polygons/carto-polygons.ts
@@ -1,7 +1,5 @@
-/* eslint-disable @typescript-eslint/camelcase */
 import { GeneratorStyles, Group } from '../../types'
 import { Type, CartoPolygonsGeneratorConfig } from '../types'
-import { isConfigVisible } from '../utils'
 import layersDirectory from './carto-polygons-layers'
 
 export const CARTO_FISHING_MAP_API = 'https://carto.globalfishingwatch.org/user/admin/api/v1/map'
@@ -65,7 +63,7 @@ class CartoPolygonsGenerator {
           })
           const tiles = [`${CARTO_FISHING_MAP_API}/${layergroupid}/{z}/{x}/{y}.mvt`]
           this.tilesCacheByid[id] = tiles
-          return this.getStyle(config)
+          return { style: this.getStyle(config), config }
         } catch (e) {
           throw e
         }
@@ -86,8 +84,7 @@ class CartoPolygonsGenerator {
 
     const layerData = (layersDirectory as any)[cartoTableId] || config
     const layers: any = layerData.layers.map((glLayer: any) => {
-      const visibility = isConfigVisible(config)
-      const layout = glLayer.layout ? { ...glLayer.layout, visibility } : { visibility }
+      const layout = glLayer.layout ? { ...glLayer.layout } : {}
       const paint: any = {}
       const hasSelectedFeatures = config.selectedFeatures?.values?.length
       // TODO: make this dynamic