+++ packages/layer-composer/src/generators/types.ts
@@ -31,8 +31,11 @@ export interface GlobalGeneratorConfig {
   start: string
   end: string
   zoom: number
-  zoomLoadLevel?: number
   token?: string
+}
+
+export interface GlobalGeneratorConfigExtended extends GlobalGeneratorConfig {
+  zoomLoadLevel: number
 }
 
 export type AnyData = FeatureCollection | Segment[] | RawEvent[] | Ruler[] | null
@@ -61,7 +64,7 @@ export interface GeneratorConfig {
  * This is the union of GeneratorConfig <T> with GlobalGeneratorConfig, which allows access to both
  * generator config params and global config params, at the generator level
  */
-export type MergedGeneratorConfig<T> = T & GlobalGeneratorConfig
+export type MergedGeneratorConfig<T> = T & GlobalGeneratorConfigExtended
 
 /**
  * A solid color background layer
@@ -266,11 +269,12 @@ export interface HeatmapAnimatedGeneratorConfig extends GeneratorConfig {
   sublayers: HeatmapAnimatedGeneratorSublayer[]
   mode?: HeatmapAnimatedMode
   tilesAPI?: string
+  breaksAPI?: string
   maxZoom?: number
   debug?: boolean
   debugLabels?: boolean
-  tilesetsStart?: string
-  tilesetsEnd?: string
+  datasetsStart?: string
+  datasetsEnd?: string
   interactive?: boolean
   staticStart?: string
   staticEnd?: string
@@ -355,29 +359,29 @@ export interface HeatmapAnimatedGeneratorSublayer {
 }
 
 // ---- Heatmap Generator color ramps types
-export type ColorRampsIds =
-  | 'fishing'
-  | 'presence'
-  | 'reception'
-  | 'bivariate'
+export type ColorRampId =
   | 'teal'
-  | 'teal_toWhite'
   | 'magenta'
-  | 'magenta_toWhite'
   | 'lilac'
-  | 'lilac_toWhite'
   | 'salmon'
-  | 'salmon_toWhite'
   | 'sky'
-  | 'sky_toWhite'
   | 'red'
-  | 'red_toWhite'
   | 'yellow'
-  | 'yellow_toWhite'
   | 'green'
-  | 'green_toWhite'
   | 'orange'
+
+export type ColorRampWhiteId =
+  | 'teal_toWhite'
+  | 'magenta_toWhite'
+  | 'lilac_toWhite'
+  | 'salmon_toWhite'
+  | 'sky_toWhite'
+  | 'red_toWhite'
+  | 'yellow_toWhite'
+  | 'green_toWhite'
   | 'orange_toWhite'
+
+export type ColorRampsIds = ColorRampId | ColorRampWhiteId
 
 export enum HeatmapAnimatedMode {
   // Pick sublayer with highest value and place across this sublayer's color ramp. Works with 0 - n sublayers